Convert type tests for saga data properties to component test
